refactor cfg module
